Change the parser code like a G to allow multiple different io flags in a sigle command
The parser would have a global flag for append operations, really? If I have a > and >> in the same command I will have the flag as append for all of the redirects.
That bothered me so I modified the parser to store a flag for all the words, this way I can check if each file is in regular or append mode.